requirements.txt
astroid==2.4.2
Django==2.2.10
django-crispy-forms==1.10.0
django-filter==2.4.0
djangorestframework==3.12.2
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
